Restructuring of the framework (Squash commit)
The old structure seemed a bit too complex and was just not
understandable after coming back to it after some time.
Most of the tests were removed but will be readded.
The frontend is useless but that will be changed later anyways.
A lot of code was changed in this commit. The Framework interfaces
were adjusted and Services introduced for internal use.
The api package contains everything that is supposed to be used
from outside.